chore: Move settings pages to parent scope to escape flow editor view
